requirements.txt
Django==2.2.5
MarkupSafe==1.1.1
pytz==2019.2
sqlparse==0.3.0
djangorestframework
djangorestframework-jwt
django-filter
django-rest-auth
django-allauth
requests